Added second copy of submenu for More… dropdown
The changes through line 5 are all just formatting.

Lines 18-34 output a second copy of the submenu to fill the dropdown panel.

#more and .menu-more are new and don’t have any CSS attach to them yet.

The dropdown-related items on lines 19 and 20 are taken from the user dropdown.

The id’s and classes on lines 26 and 27 are taken from lines 10 and 11. This double usage may need to be revised. At the moment, it’s just in as a placeholder/prototype.